split the cpu core
